+++ src/Domain/Entity/Mill/Pipe.cs
@@ -21,7 +21,7 @@ namespace Prizm.Domain.Entity.Mill
 
         public virtual string Mill { get; set; }
         public virtual int Diameter { get; set; }
-        public virtual int WallThickness { get; set; }
+        public virtual float WallThickness { get; set; }
 
         public virtual float Weight { get; set; }
